requirements.txt
fastapi
uvicorn
httpx
pytest-dotenv
pytest-asyncio
respx
pydantic>=2,<3.0.0
black
isort
pylint
autoflake
